+++ .eslintrc
@@ -1,7 +1,3 @@
 {
-    "extends": ["standard",
-      "plugin:react/recommended"],
-    "env": {
-        "jest": true
-    }
+    "extends": "standard"
 }
